requirements.txt
python-telegram-bot>=21.0
torch>=2.0
transformers>=4.30
Pillow>=10.0
opencv-python-headless>=4.8
Flask>=3.0
flask-socketio>=5.3
python-dotenv>=1.0
